Primeras correcciones
